requirements.txt
django-heroku==0.3.1
gunicorn==20.1.0
virtualenv==20.4.3
Django==4.1.1
Pillow==8.2.0
whitenoise==6.2.0
psycopg2-binary==2.9.3
pip==22.2.2
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-widget-tweaks==1.4.12
python-decouple==3.6
django-jazzmin==2.5.0
django-phonenumber-field-7.0.0